requirements.txt
pymupdf
fastapi
uvicorn
sentence-transformers
fastapi
uvicorn
pymupdf
chromadb
sentence-transformers
python-dotenv
langchain
langchain-community
langchain-text-splitters
google-generativeai